feat: type aliases
It is now possible to add type aliases to the `TypeGenerator`:

    gen.addAlias('FromMyType', 'ToYourType');

The implication is that every `$ref` to `FromMyType` will be resolved as `ToYourType` instead.

The destination type must either has a definition (added via `definitions` or `addDefinition()`)  or it can also be a custom type added via `emitCustomType()`.

Misc:

- Renamed `addType()` to `emitType()` (soft-deprecation).
- Renamed `addCode()` to `emitCustomType()` (soft-deprecation)
- Added `addDefinition(typeName, schema)`
- Support calling `emitType(name)` without a schema - this will look up the definition of the
  type automatically.

Required in order to implement awslabs/cdk8s#370