feat: generate toJson() functions for structs
Since the type generator converts property names to camel case (so they are JSII compliant), users need to be able to convert a data struct back to schema format. Technically this conversion is lossy, so in order to facilitate that back-conversion we now automatically generate a function called `xxx_toJson` for each struct which recursively converts the data back.

This also fixes #343