fix: unable to reference multiple structs that share a dotwise base name (#509)
As mentioned in this [comment](https://github.com/cdk8s-team/cdk8s-cli/pull/54#issuecomment-899598963), the underlying problem was that when types are created from references, the type name used to only be the last element of a possibly dotwise namespace.

This reduced references like `io.k8s.v1beta.Foo` and `io.k8s.v1.Foo` to resolve to the same type name (i.e `Foo`), causing only the first of which to be emitted and referenced to. 

This PR applies a normalization logic to those references. By default, dotwise namespaces are normalized by pascal casing each element and joining them together, just as we already do for property types:

https://github.com/cdklabs/json2jsii/blob/0ddf8e1448d891c5f8d9173592372f79c6698651/src/type-generator.ts#L484-L486

This will result in interfaces named `IoK8Sv1Beta1Foo` - which might be a little too verbose. To that end, we also provide a way to pass a custom ref normalizer.

Fixes https://github.com/cdk8s-team/cdk8s-cli/issues/45