fix: unions with multiple of the same type produce invalid typescript (#1799)
Previously a schema like this:

```
{
  properties: {
    foo: {
      oneOf: [
        {
          type: 'string',
          enum: [
            'tab',
          ],
        },
        {
          type: 'string',
          enum: [
            'space',
          ],
        },
      ],
    },
  },
}
```

was producing invalid typescript code (note the doubling of
`fromString`):
```ts
export interface TestType {
  readonly foo?: FqnOfTestTypeFoo;
}
export class FqnOfTestTypeFoo {
  public static fromString(value: string): Foo {
    return new FqnOfTestTypeFoo(value);
  }
  public static fromString(value: string): Foo {
    return new FqnOfTestTypeFoo(value);
  }
  private constructor(public readonly value: string | string) {
  }
}
```

instead we now create valid code:


```ts
export interface TestType {
  readonly foo?: FqnOfTestTypeFoo;
}
export enum FqnOfTestTypeFoo {
  /** tab */
  TAB = "tab",
  /** space */
  SPACE = "space",
}
```

---

Introduces the new concept of transformers. Transformers are either
mandatory or opt-in (not yet available) conversions of JSONSchema
definitions into a simplified, but semantically equivalent form. This
reduces the need for the type generator to be aware of edge cases, but
also gives the user more control over how to handle edge cases.

**Mandatory transformers** are such ones necessary to generate valid
code. An example is available in this PR: A union schema with multiple
of the same type produces invalid code, thus reducing the union to
elements of unique types is a mandatory transform.

**Opt-in transformers** are such ones that improve the generated code,
but are either not backwards compatible or carry the risk of it. For
example a union schema with a single element could be "hoisted" up to
the top level to avoid the union all together. This produces cleaner
code. However the schema author might have use a union to convey their
intention of adding further elements into the union in future, which
would than cause a breaking change in the generated code.

---

### Ask Yourself

- [x] Have you reviewed the [contribution
guide](https://github.com/cdklabs/json2jsii/blob/main/CONTRIBUTING.md)?
- [x] Have you reviewed the [breaking changes
guide](https://github.com/cdklabs/json2jsii/blob/main/CONTRIBUTING.md#breaking-changes)?

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>